Introduces base classes for test cases and test case builders
