+++ Test/test_utils/TestCaseBuilderBase.java
@@ -1,0 +1,20 @@
+package test_utils;
+
+import calculator.Vehicle;
+
+public class TestCaseBuilderBase {
+    String nameHeader;
+    String nameTail;
+    Vehicle actualVehicle;
+    int expected;
+
+    String name() {
+        if (nameHeader == null) {
+            return nameTail;
+        } else if (nameTail == null) {
+            return nameHeader;
+        } else {
+            return nameHeader + ": " + nameTail;
+        }
+    }
+}
