+++ Test/test_utils/TestCaseWithMultipleDates.java
@@ -4,21 +4,11 @@ import calculator.Vehicle;
 
 import java.util.Date;
 
-public class TestCaseWithMultipleDates {
-    public final String name;
-    public final Vehicle actualVehicle;
+public class TestCaseWithMultipleDates extends TestCaseBase {
     public final Date[] actualTimes;
-    public final int expected;
 
     public TestCaseWithMultipleDates(String name, Date[] actualTimes, Vehicle actualVehicle, int expected) {
-        this.name = name;
+        super(expected, actualVehicle, name);
         this.actualTimes = actualTimes;
-        this.actualVehicle = actualVehicle;
-        this.expected = expected;
-    }
-
-    @Override
-    public String toString() {
-        return name;
     }
 }
